requirements.txt
antiorm==1.2.1
asgiref==3.7.2
base64io==1.0.3
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
chartkick==1.0.1
click==8.1.6
db==0.1.1
Django==4.2.4
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
WTForms==3.0.1
